refactor: 우승자 찾기 위치 변경
